+++ common/lib/src/realm_common_base.dart
@@ -105,3 +105,11 @@ class Backlink {
   final Symbol fieldName;
   const Backlink(this.fieldName);
 }
+
+/// @nodoc
+class Tuple<T1, T2> {
+  T1 item1;
+  T2 item2;
+
+  Tuple(this.item1, this.item2);
+}