+++ lib/src/realm_class.dart
@@ -83,7 +83,17 @@ export "configuration.dart"
 export 'credentials.dart' show Credentials, AuthProviderType, EmailPasswordAuthProvider;
 export 'list.dart' show RealmList, RealmListOfObject, RealmListChanges;
 export 'realm_object.dart'
-    show RealmEntity, RealmException, UserCallbackException, RealmObject, RealmObjectBase, EmbeddedObject, RealmObjectChanges, DynamicRealmObject;
+    show
+        RealmEntity,
+        RealmException,
+        UserCallbackException,
+        RealmObject,
+        RealmObjectBase,
+        EmbeddedObject,
+        EmbeddedObjectExtension,
+        RealmObjectChanges,
+        DynamicRealmObject;
+
 export 'realm_property.dart';
 export 'results.dart' show RealmResults, RealmResultsChanges, RealmResultsOfObject;
 export 'session.dart' show Session, SessionState, ConnectionState, ProgressDirection, ProgressMode, SyncProgress, ConnectionStateChange;
@@ -756,6 +766,20 @@ class RealmMetadata {
     }
 
     return metadata;
+  }
+
+  Tuple<Type, RealmObjectMetadata> getByClassKey(int key) {
+    final type = _typeMap.entries.firstWhereOrNull((e) => e.value.classKey == key);
+    if (type != null) {
+      return Tuple(type.key, type.value);
+    }
+
+    final metadata = _stringMap.values.firstWhereOrNull((e) => e.classKey == key);
+    if (metadata != null) {
+      return Tuple(RealmObjectBase, metadata);
+    }
+
+    throw RealmError("Object with classKey $key not found in the current Realm's schema.");
   }
 }